+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/calendar-filter/CalendarFilterDefault.test.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import CalendarFilter from '../../../../CalendarFilter';
+
+const CalendarFilterDefault = () => (
+  <CalendarFilter selectedDate="2020-04-04" />
+);
+
+export default CalendarFilterDefault;