+++ packages/terra-hookshot/src/terra-dev-site/doc/example/HookshotExample.jsx
@@ -54,7 +54,7 @@ class HookshotStandard extends React.Component {
     super(props);
     this.handleButtonClick = this.handleButtonClick.bind(this);
     this.handleRequestClose = this.handleRequestClose.bind(this);
-    this.handleAttachementBehaviorChange = this.handleAttachementBehaviorChange.bind(this);
+    this.handleAttachmentBehaviorChange = this.handleAttachmentBehaviorChange.bind(this);
     this.handleContentAttachmentChange = this.handleContentAttachmentChange.bind(this);
     this.handleTargetAttachmentChange = this.handleTargetAttachmentChange.bind(this);
     this.handleInputChange = this.handleInputChange.bind(this);
@@ -75,7 +75,7 @@ class HookshotStandard extends React.Component {
     this.setState({ isOpen: false });
   }
 
-  handleAttachementBehaviorChange(event) {
+  handleAttachmentBehaviorChange(event) {
     this.setState({ hookshotAttachmentBehavior: event.target.value });
   }
 
@@ -88,8 +88,9 @@ class HookshotStandard extends React.Component {
   }
 
   handleInputChange(event) {
-    const value = Number.parseFloat(event.target.value);
-    this.setState({ [event.target.name]: value });
+    if (event.target.value !== '') {
+      this.setState({ [event.target.name]: Number.parseFloat(event.target.value) });
+    }
   }
 
   render() {
@@ -112,7 +113,7 @@ class HookshotStandard extends React.Component {
           id={getId('hookshotAttachmentBehavior')}
           name="hookshotAttachmentBehavior"
           value={this.state.hookshotAttachmentBehavior}
-          onChange={this.handleAttachementBehaviorChange}
+          onChange={this.handleAttachmentBehaviorChange}
           className={cx('select-wrapper')}
         >
           {generateOptions(Hookshot.attachmentBehaviors)}